feat: 친구 검색 및 추가 기능 추가

+++ src/main/java/eolebeole/bemealmap/domain/entity/User.java
@@ -16,7 +16,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
+    private int userId;
 
     @Column
     private String email;
